tracepoint: Add a first working capture of tracepoints
This commit introduces the base object of obs, the observer. One attaches
things he wants to observe to an observer, like Tracepoints, and read events
from it.

Then, tracepoints. This object corresponds to a kernel tracepoint. The API
looks like (omitting error handing):

	import "obs"

	exec := obs.NewTracepoint("sched:sched_process_exec")

	observer := obs.NewObserver()
        observer.AddTracepoint(exec)
        observer.Open()

        for {
        	event, _ := observer.ReadEvent()
		/* do something with the event */
        }

This will listen for sched_process_exec tracepoints and return an event object
each time the tracepoint is triggered.